split nativeapp intg tests into files
