+++ tests_integration/nativeapp/test_init_run.py
@@ -1,0 +1,406 @@
+import os
+import uuid
+
+from snowflake.cli.api.project.util import generate_user_env
+from snowflake.cli.api.secure_path import SecurePath
+from snowflake.cli.plugins.nativeapp.init import OFFICIAL_TEMPLATES_GITHUB_URL
+
+from tests.project.fixtures import *
+from tests_integration.test_utils import (
+    pushd,
+    contains_row_with,
+    not_contains_row_with,
+    row_from_snowflake_session,
+)
+
+USER_NAME = f"user_{uuid.uuid4().hex}"
+TEST_ENV = generate_user_env(USER_NAME)
+
+
+# Tests a simple flow of initiating a new project, executing snow app run and teardown, all with distribution=internal
+@pytest.mark.integration
+def test_nativeapp_init_run_without_modifications(
+    runner,
+    snowflake_session,
+    temporary_working_directory,
+):
+    project_name = "myapp"
+    result = runner.invoke_json(
+        ["app", "init", project_name],
+        env=TEST_ENV,
+    )
+    assert result.exit_code == 0
+
+    with pushd(Path(os.getcwd(), project_name)):
+        result = runner.invoke_with_connection_json(
+            ["app", "run"],
+            env=TEST_ENV,
+        )
+        assert result.exit_code == 0
+
+        try:
+            # app + package exist
+            package_name = f"{project_name}_pkg_{USER_NAME}".upper()
+            app_name = f"{project_name}_{USER_NAME}".upper()
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"show application packages like '{package_name}'",
+                    )
+                ),
+                dict(name=package_name),
+            )
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"show applications like '{app_name}'",
+                    )
+                ),
+                dict(name=app_name),
+            )
+
+            # make sure we always delete the app
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+        finally:
+            # teardown is idempotent, so we can execute it again with no ill effects
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown", "--force"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+
+# Tests a simple flow of an existing project, but executing snow app run and teardown, all with distribution=internal
+@pytest.mark.integration
+@pytest.mark.parametrize("project_definition_files", ["integration"], indirect=True)
+def test_nativeapp_run_existing(
+    runner,
+    snowflake_session,
+    project_definition_files: List[Path],
+):
+    project_name = "integration"
+    project_dir = project_definition_files[0].parent
+    with pushd(project_dir):
+        result = runner.invoke_with_connection_json(
+            ["app", "run"],
+            env=TEST_ENV,
+        )
+        assert result.exit_code == 0
+
+        try:
+            # app + package exist
+            package_name = f"{project_name}_pkg_{USER_NAME}".upper()
+            app_name = f"{project_name}_{USER_NAME}".upper()
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"show application packages like '{package_name}'",
+                    )
+                ),
+                dict(name=package_name),
+            )
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"show applications like '{app_name}'"
+                    )
+                ),
+                dict(name=app_name),
+            )
+
+            # sanity checks
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"select count(*) from {app_name}.core.shared_view"
+                    )
+                ),
+                {"COUNT(*)": 1},
+            )
+            test_string = "TEST STRING"
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"call {app_name}.core.echo('{test_string}')"
+                    )
+                ),
+                {"ECHO": test_string},
+            )
+
+            # make sure we always delete the app
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+        finally:
+            # teardown is idempotent, so we can execute it again with no ill effects
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown", "--force"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+
+# Tests a simple flow of initiating a project, executing snow app run and teardown, all with distribution=internal
+@pytest.mark.integration
+def test_nativeapp_init_run_handles_spaces(
+    runner,
+    snowflake_session,
+    temporary_working_directory,
+):
+    project_name = "myapp"
+    project_dir = "app root"
+    result = runner.invoke_json(
+        ["app", "init", project_dir, "--name", project_name],
+        env=TEST_ENV,
+    )
+    assert result.exit_code == 0
+
+    with pushd(Path(os.getcwd(), project_dir)):
+        result = runner.invoke_with_connection_json(
+            ["app", "run"],
+            env=TEST_ENV,
+        )
+        assert result.exit_code == 0
+
+        try:
+            # app + package exist
+            package_name = f"{project_name}_pkg_{USER_NAME}".upper()
+            app_name = f"{project_name}_{USER_NAME}".upper()
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"show application packages like '{package_name}'",
+                    )
+                ),
+                dict(name=package_name),
+            )
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"show applications like '{app_name}'",
+                    )
+                ),
+                dict(name=app_name),
+            )
+
+            # make sure we always delete the app
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+        finally:
+            # teardown is idempotent, so we can execute it again with no ill effects
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown", "--force"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+
+# Tests a simple flow of an existing project, but executing snow app run and teardown, all with distribution=external
+@pytest.mark.integration
+@pytest.mark.parametrize(
+    "project_definition_files", ["integration_external"], indirect=True
+)
+def test_nativeapp_run_existing_w_external(
+    runner,
+    snowflake_session,
+    project_definition_files: List[Path],
+):
+    project_name = "integration_external"
+    project_dir = project_definition_files[0].parent
+    with pushd(project_dir):
+        result = runner.invoke_with_connection_json(
+            ["app", "run"],
+            env=TEST_ENV,
+        )
+        assert result.exit_code == 0
+
+        try:
+            # app + package exist
+            package_name = f"{project_name}_pkg_{USER_NAME}".upper()
+            app_name = f"{project_name}_{USER_NAME}".upper()
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"show application packages like '{package_name}'",
+                    )
+                ),
+                dict(name=package_name),
+            )
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"show applications like '{app_name}'"
+                    )
+                ),
+                dict(name=app_name),
+            )
+
+            # app package contains distribution=external
+            expect = row_from_snowflake_session(
+                snowflake_session.execute_string(
+                    f"desc application package {package_name}"
+                )
+            )
+            assert contains_row_with(
+                expect, {"property": "name", "value": package_name}
+            )
+            assert contains_row_with(
+                expect, {"property": "distribution", "value": "EXTERNAL"}
+            )
+
+            # sanity checks
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"select count(*) from {app_name}.core.shared_view"
+                    )
+                ),
+                {"COUNT(*)": 1},
+            )
+            test_string = "TEST STRING"
+            assert contains_row_with(
+                row_from_snowflake_session(
+                    snowflake_session.execute_string(
+                        f"call {app_name}.core.echo('{test_string}')"
+                    )
+                ),
+                {"ECHO": test_string},
+            )
+
+            # make sure we always delete the app, --force required for external distribution
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown", "--force"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+            expect = snowflake_session.execute_string(
+                f"show applications like '{app_name}'"
+            )
+            assert not_contains_row_with(
+                row_from_snowflake_session(expect), {"name": app_name}
+            )
+
+            expect = snowflake_session.execute_string(
+                f"show application packages like '{package_name}'"
+            )
+            assert not_contains_row_with(
+                row_from_snowflake_session(expect), {"name": package_name}
+            )
+
+        finally:
+            # teardown is idempotent, so we can execute it again with no ill effects
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown", "--force"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+
+# Verifies that running "app run" after "app deploy" upgrades the app
+@pytest.mark.integration
+def test_nativeapp_run_after_deploy(
+    runner,
+    temporary_working_directory,
+):
+    project_name = "myapp"
+    app_name = f"{project_name}_{USER_NAME}"
+    stage_fqn = f"{project_name}_pkg_{USER_NAME}.app_src.stage"
+
+    result = runner.invoke_json(
+        ["app", "init", project_name],
+        env=TEST_ENV,
+    )
+    assert result.exit_code == 0
+
+    with pushd(Path(os.getcwd(), project_name)):
+        try:
+            # Run #1
+            result = runner.invoke_with_connection_json(
+                ["app", "run"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+            # Make a change & deploy
+            with open("app/README.md", "a") as file:
+                file.write("### Test")
+            result = runner.invoke_with_connection_json(
+                ["app", "deploy"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+            # Run #2
+            result = runner.invoke_with_connection_json(
+                ["app", "run", "--debug"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+            assert (
+                f"alter application {app_name} upgrade using @{stage_fqn}"
+                in result.output
+            )
+
+        finally:
+            result = runner.invoke_with_connection_json(
+                ["app", "teardown", "--force"],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+
+
+# Tests initialization of a project from a repo with a single template
+@pytest.mark.integration
+def test_nativeapp_init_from_repo_with_single_template(
+    runner,
+    snowflake_session,
+    temporary_working_directory,
+):
+    from git import Repo
+    from git import rmtree as git_rmtree
+
+    with SecurePath.temporary_directory() as all_templates_local_repo_path:
+        # prepare a local repository with only one template (basic)
+        all_templates_repo = Repo.clone_from(
+            url=OFFICIAL_TEMPLATES_GITHUB_URL,
+            to_path=all_templates_local_repo_path.path,
+            filter=["tree:0"],
+            depth=1,
+        )
+        all_templates_repo.close()
+        git_rmtree((all_templates_local_repo_path / ".git").path)
+
+        single_template_repo_path = all_templates_local_repo_path / "basic"
+        single_template_repo = Repo.init(single_template_repo_path.path)
+        single_template_repo.index.add(["**/*", "*", ".gitignore"])
+        single_template_repo.index.commit("initial commit")
+
+        # confirm that no error is thrown when initializing a project from a repo with a single template
+        project_name = "myapp"
+        try:
+            result = runner.invoke_json(
+                [
+                    "app",
+                    "init",
+                    "--template-repo",
+                    f"file://{single_template_repo_path.path}",
+                    project_name,
+                ],
+                env=TEST_ENV,
+            )
+            assert result.exit_code == 0
+        finally:
+            single_template_repo.close()
